pyproject.toml
[tool.poetry]
name = "gitea.py"
version = "0.1.0"
description = "An API wrapper for Gitea written in Python"
authors = ["Joakim Hellsén <tlovinator@gmail.com>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.23.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
python-dotenv = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
